requirements.txt
Django==3.2.20
djangorestframework==3.14.0
mysqlclient==2.1.1
requests>=2.26,<=2.28